feat: grab funds tooltip
